@@ -1,5 +1,5 @@
 @page "/read/{slug}"
-@model Dice_Driven_Stories.Pages.ArticleModel
+@model Swrith.Pages.ArticleModel
 @{
     ViewData["Title"] = @Model.PostTitle;
 }
