@@ -1,5 +1,5 @@
 ﻿@page
-@model Dice_Driven_Stories.Pages.IndexModel
+@model Swrith.Pages.IndexModel
 @{
     ViewData["Title"] = "Home page";
 }
