@@ -1,5 +1,5 @@
 @page "/read/contains"
-@model Dice_Driven_Stories.Pages.SearchModel
+@model Swrith.Pages.SearchModel
 @{
     ViewData["Title"] = "Search";
 }
