@@ -3,14 +3,13 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"] - Dice Driven Stories</title>
+    <title>@ViewData["Title"] - Swrith</title>
     <meta name="author" content="Hainezo B."/>
     <meta name="description" content="A blog of different stories driven by dice roll."/>
-    <meta name="keywords" content="RDS, Dice, driven, stories, blog, personnal, enzo, bruni, article, articles, dice, die, fantasy, tolkien, horror, story, writer, writing, words, hainezo, book, books, english, random"/>
+    <meta name="keywords" content="short, stories, blog, personnal, enzo, bruni, article, articles, fantasy, story, writer, writing, words, hainezo, book, books, english, random"/>
     <meta  property="og:type" content="article">
-    <meta  property="og:site_name" content="Dice Driven Stories">
+    <meta  property="og:site_name" content="Swrith">
     <meta  property="og:image" content="~/bookicon.ico">
-    <meta name="google-site-verification" content="4Y_x2_iY_Ho67saxELMzs5klxbPl_ZDqr45MpUcWSH0" />
     <environment include="Development">
         <link rel="stylesheet" href="~/css/site.css" />
     </environment>
