@@ -1,6 +1,6 @@
 <nav>
     <h1>
-        <a asp-page="Index">Dice Driven Stories</a>
+        <a asp-page="Index">Swrith</a>
     </h1>
     
     <form method="get" action="/read/contains">
