@@ -1,3 +1,3 @@
-@using Dice_Driven_Stories
-@namespace Dice_Driven_Stories.Pages
+@using Swrith
+@namespace Swrith.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
